Bug fix for MLJ data input process
Fixing the bug that produces the following error:
```
The scitype of `X`, in `machine(model, X, ...)` is incompatible with `model=MultinomialClassifier`:
scitype(X) = AbstractMatrix{ScientificTypesBase.Continuous}
input_scitype(model) = ScientificTypesBase.Table{<:AbstractVector{<:ScientificTypesBase.Continuous}}.
```
This is due to the input data `X` being an array instead of a data frame. The fix for this
is to convert all the `X` into data frames.